deep-insight-web/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
boto3>=1.38.0
python-dotenv>=1.1.0
python-multipart>=0.0.22
PyJWT>=2.9.0
cryptography>=43.0.0
